redis-worker/requirements.txt
# requirements.txt for redis-worker
redis
numpy
Pillow